requirements.txt
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
blinker==1.8.2
certifi==2024.7.4
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.2
Flask==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
openai==1.37.1
openapi==2.0.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
referencing==0.35.1
rpds-py==0.19.1
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.2
Werkzeug==3.0.3